Assignment-2
In this example, we create an array called "fruits" and assign it the specified fruits using square brackets []. The fruits.length property gives us the total number of elements in the array. We use console.log() to display the result, concatenating it with the string "Total number of fruits: ". When you run this code, it will output the total number of fruits in the array, which in this case is 5.